Create 0MQ applications enabling enabling communication between robobee & queenbee

+++ tests/services_test/robo_bee_svc_test/src/robo_bee_svc_test.cpp
@@ -21,8 +21,8 @@ TEST(RoboBeeSvc, TestCase1Waypoint) {
 
   kpsr::mem::MemEnv environment;
 
-  float imgWidth = 320;
-  float imgHeight = 240;
+  int imgWidth = 320;
+  int imgHeight = 240;
 
   mls::RoboBeeSvc roboBeeSvc(&environment, imageDataProvider.getPublisher(),
                              waypointProvider.getSubscriber(), imgWidth,
